refactor: Move exception classes to 'error' package
