+++ domain/src/main/java/com/london/domain/error/exceptions.kt
@@ -1,4 +1,4 @@
-package com.london.domain
+package com.london.domain.error
 
 class MovieSearchFailedException(message: String = "Failed to search for movies.") : Exception(message)
 class TvShowSearchFailedException(message: String = "Failed to search for TV shows.") : Exception(message)
