+++ domain/src/test/kotlin/com/london/domain/usecase/GetActorsUseCaseTest.kt
@@ -1,9 +1,9 @@
 package com.london.domain.usecase
 
 import com.google.common.truth.Truth.assertThat
-import com.london.domain.ActorSearchFailedException
 import com.london.domain.entity.Actor
 import com.london.domain.entity.PagedFetchResponse
+import com.london.domain.error.ActorSearchFailedException
 import com.london.domain.repository.SearchRepository
 import io.mockk.coEvery
 import io.mockk.mockk
