+++ domain/src/test/kotlin/com/london/domain/usecase/GetMovieCastUseCaseTest.kt
@@ -1,8 +1,8 @@
 package com.london.domain.usecase
 
 import com.google.common.truth.Truth.assertThat
-import com.london.domain.GetCastByIdFailedException
 import com.london.domain.entity.Actor
+import com.london.domain.error.GetCastByIdFailedException
 import com.london.domain.repository.MovieDetailsRepository
 import io.mockk.coEvery
 import io.mockk.coVerify
