+++ domain/src/test/kotlin/com/london/domain/usecase/GetSimilarMoviesUseCaseTest.kt
@@ -1,7 +1,7 @@
 package com.london.domain.usecase
 
 import com.google.common.truth.Truth.assertThat
-import com.london.domain.GetMovieCastFailedException
+import com.london.domain.error.GetMovieCastFailedException
 import com.london.domain.entity.moviedatails.SimilarMovie
 import com.london.domain.repository.MovieDetailsRepository
 import io.mockk.coEvery
