+++ domain/src/test/kotlin/com/london/domain/usecase/GetUpComingMoviesByCategoryUseCaseTest.kt
@@ -1,7 +1,7 @@
 package com.london.domain.usecase
 
 import com.google.common.truth.Truth.assertThat
-import com.london.domain.MovieSearchFailedException
+import com.london.domain.error.MovieSearchFailedException
 import com.london.domain.entity.Movie
 import com.london.domain.entity.PagedFetchResponse
 import com.london.domain.repository.SearchRepository
